# [Security Scans] Adding performance test for auto scans