--- a/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
+++ b/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
@@ -18,7 +18,7 @@ import * as fs from 'fs'
 import fileSystem from '../../../shared/fs/fs'
 import * as os from 'os'
 import * as path from 'path'
-import { assertTelemetry, closeAllEditors } from '../../testUtil'
+import { assertTelemetry, closeAllEditors, createTestWorkspaceFolder, toFile } from '../../testUtil'
 import { stub } from '../../utilities/stubber'
 import { AWSError, HttpResponse } from 'aws-sdk'
 import { getTestWindow } from '../../shared/vscode/window'
@@ -458,9 +458,13 @@ describe('startSecurityScanPerformanceTest', function () {
         extensionContext = await FakeExtensionContext.create()
         mockSecurityPanelViewProvider = new SecurityPanelViewProvider(extensionContext)
 
+        const folder = await createTestWorkspaceFolder()
+        const mockFilePath = path.join(folder.uri.fsPath, 'app.py')
+        await toFile('hello_world', mockFilePath)
+        /*
         // Create a temp directory
         const tempDirectory = await fs.promises.mkdtemp(path.join(os.tmpdir(), 'temp-'))
-
+       
         // Create a dummy 200KB file as app.py in the temp directory
         const mockFilePath = path.join(tempDirectory, 'hello_world', 'app.py')
         await fs.promises.mkdir(path.dirname(mockFilePath), { recursive: true })
@@ -468,7 +472,7 @@ describe('startSecurityScanPerformanceTest', function () {
         // Create a sample text content with 199 KB size
         const mockContent = 'print("Hello, World!")'.repeat((199 * 1024) / 22)
         await fileSystem.writeFile(mockFilePath, mockContent)
-
+*/
         appCodePath = mockFilePath
         editor = await openTestFile(appCodePath)
         await model.CodeScansState.instance.setScansEnabled(false)