--- a/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
+++ b/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
@@ -528,7 +528,8 @@ describe('startSecurityScanPerformanceTest', function () {
                 User CPU Usage: 0.058255 sec and System CPU Usage: 0.004867 sec
          */
 
-        assert(EndScanCpuUsageByUser > 0, `User CPU usage difference should be greater than 0 but actual is ${EndScanCpuUsageByUser}`)
+        assert(EndScanCpuUsageByUser > 0,
+            `User CPU usage difference should be greater than 0 but actual is ${EndScanCpuUsageByUser}`)
         // These limits are considered from local mac but may vary with machine, OS etc factors
         assert(
             EndScanMemoryUsageInMB < 400,