--- a/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
+++ b/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
@@ -451,7 +451,6 @@ describe('startSecurityScan', function () {
 })
 
 describe('startSecurityScanPerformanceTest', function () {
-
     beforeEach(async function () {
         extensionContext = await FakeExtensionContext.create()
         mockSecurityPanelViewProvider = new SecurityPanelViewProvider(extensionContext)
@@ -519,7 +518,7 @@ describe('startSecurityScanPerformanceTest', function () {
         const cpuUsagePercentage = ((EndScanCpuUsageByUser + EndScanCpuUsageBySystem) / elapsedSeconds) * 100
 
         // These limits are considered from local mac but may vary with machine, OS etc factors
-        assert(cpuUsagePercentage > 50, `User CPU usage should be less than 50% of total CPU%`)
+        assert(cpuUsagePercentage < 50, `Expected CPU usage should be less than 50% of total CPU%, actual CPU usage is ${cpuUsagePercentage}`)
 
         assert(
             EndScanMemoryUsageInMB < 300,