## Problem
- Security scans does not have performance tests for file scan.
## Solution
- Added performance test for file scans with payload size of 200KB(Max Payload size).
- Observations in local machine(Mac)

**Input:**

> Payload Size is 198.99 KB

**Result:**

| Machine  | CPU Usage % | Memory Usage in MB  | 
| ------------- | ------------- | ------------- | 
| Local Machine(Mac OS)  | 25.34  | 239.0625  |
| MacOS 16.x minimum  | 23.08  | 228.32813  |
| MacOS 16.x stable  | 19.97  | 209.23438  |
| MacOS 16.x insiders  | 22.45  | 216.48438 |
| Windows 16.x stable  | 39.98  | 206.50391  |
| Windows 16.x insiders  | -  | 214.00391  |


> Memory Usage for performing a file scan : 203.75390625 MB
> Start of a Scan:
> User CPU Usage: 21.446874 sec and System CPU Usage: 7.044477 sec
> Difference of CPU Usage at End of a Scan:
> User CPU Usage: 0.058255 sec and System CPU Usage: 0.004867 sec

This depends on machine, os and other factors, currently gave a limit of 400MB max memory usage but its volatile.



## License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.