@hayemaxi 2024-08-06
Interesting, windows seems to use much more CPU (but we would need a larger sample size to be more confident about that). Maybe we should consider having separate test threholds for each OS, e.g. windows: 50%, mac: 35%, linux: ?%